move whitelisting checks into WhitelistingSettings
