+++ src/test/resources/legacy/securityconfig_v6/migration/whitelist.yml
